task: categorize techs and tools and  make contacts responsive on small screen
