+++ src/app/projects/components/menu/Menu.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React, { Dispatch, SetStateAction, useContext } from "react";
+import React, { Dispatch, SetStateAction, useContext, useState } from "react";
 
 interface SectionsInterface {
   name: string;
@@ -16,7 +16,7 @@ const sections: SectionsInterface[] = [
     sectionName: "tech-stacks",
   },
   {
-    name: "Tools",
+    name: "Dev Tools",
     sectionName: "tools",
   },
   {
@@ -30,13 +30,20 @@ interface MenuProps {
 }
 
 const Menu: React.FC<MenuProps> = ({ setSelectedSection }) => {
+  const [selected, setSelected] = useState<string>("");
+
   return (
     <nav className="rounded-full bg-neutral-800 h-10 md:h-12 flex items-center justify-between gap-3 md:gap-7 px-7 md:px-12 shadow-md hover:shadow-xl dark:bg-white">
       {sections.map((section, index) => (
         <p
           key={index}
-          className="text-xs md:text-lg font-bold text-white dark:text-black cursor-pointer hover:border-y hover:border-y-white dark:hover:border-y dark:hover:border-y-black px-1"
+          className={`text-xs md:text-lg font-bold text-white dark:text-black cursor-pointer hover:border-y hover:border-y-white dark:hover:border-y dark:hover:border-y-black px-1 ${
+            section.sectionName == selected
+              ? "border-y border-y-white dark:border-y dark:border-y-black"
+              : ""
+          }`}
           onClick={() => {
+            setSelected(section.sectionName);
             setSelectedSection(section.sectionName);
           }}
         >
