+++ src/app/projects/components/sections/Tools.tsx
@@ -4,13 +4,7 @@ import React from "react";
 const Tools = () => {
   return (
     <MotionTemplate>
-      <div className="grid grid-cols-3 gap-y-2 place-items-center">
-        {Array(5)
-          .fill(0)
-          .map((_, index) => (
-            <div className="h-72 w-72 bg-black" key={index} />
-          ))}
-      </div>
+      <div className="text-white dark:text-black">dev tools</div>
     </MotionTemplate>
   );
 };
